Implements queue background service

+++ src/Brady.ScrapRunner.Mobile/Brady.ScrapRunner.Mobile/Models/TripSegmentContainerModel.cs
@@ -67,9 +67,9 @@
 
         public int? TripSegContainerWeightTare { get; set; }
 
-        public double? TripSegContainerLatitude { get; set; }
+        public int? TripSegContainerLatitude { get; set; }
 
-        public double? TripSegContainerLongitude { get; set; }
+        public int? TripSegContainerLongitude { get; set; }
 
         [MaxLength(1)]
         public string TripSegContainerLoaded { get; set; }
